+++ src/Manager/Sync/ComponentSync.php
@@ -5,11 +5,11 @@ namespace Torr\Storyblok\Manager\Sync;
 use Torr\Cli\Console\Style\TorrStyle;
 use Torr\Storyblok\Api\Data\ComponentImport;
 use Torr\Storyblok\Api\ManagementApi;
-use Torr\Storyblok\Component\Reference\ComponentsWithTags;
 use Torr\Storyblok\Exception\Api\ApiRequestException;
 use Torr\Storyblok\Exception\InvalidComponentConfigurationException;
 use Torr\Storyblok\Exception\Sync\SyncFailedException;
 use Torr\Storyblok\Manager\ComponentManager;
+use Torr\Storyblok\Manager\Sync\Filter\ResolvableComponentFilter;
 
 final class ComponentSync
 {
@@ -86,7 +86,7 @@ final class ComponentSync
 	/**
 	 * Resolves the given component config.
 	 *
-	 * This means resolving all {@see ComponentsWithTags} to their keys.
+	 * This means resolving all {@see ResolvableComponentFilter} to their keys.
 	 */
 	private function resolveComponentConfig (array $config) : array
 	{
@@ -94,11 +94,23 @@ final class ComponentSync
 
 		foreach ($config as $key => $value)
 		{
+			if ($value instanceof ResolvableComponentFilter)
+			{
+				foreach ($value->transformToManagementApiData($this->componentManager) as $nestedKey => $nestedValue)
+				{
+					$resolved[$nestedKey] = $nestedValue;
+				}
+				continue;
+			}
+
 			$resolved[$key] = match (true)
 			{
-				$value instanceof ComponentsWithTags => $this->componentManager->getComponentKeysForTags($value->tags),
 				\is_array($value) => $this->resolveComponentConfig($value),
-				default => $value,
+				\is_scalar($value) => $value,
+				default => throw new SyncFailedException(\sprintf(
+					"Invalid config value encountered: %s",
+					\get_debug_type($value),
+				)),
 			};
 		}