+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ============
 
 * (improvement) Ignore unknown bloks in `BloksField`.
+* (feature) Add `ComponentFilter` for fields to use for simpler component filtering.
+* (bc) Remove `ComponentsWithTags`, as it is replaced with `ComponentFilter`.
+* (bc) Remove ability to filter by component groups.
 
 
 3.0.0-beta.5